requirements.txt
certifi==2017.11.5
chardet==3.0.4
Django==1.11.7
django-recaptcha==1.3.1
django-recaptcha2==1.0.3
idna==2.6
pytz==2017.3
requests==2.18.4
urllib3==1.22
